move transaction processor adapter to storage
